Split structs of connectors to the dedicated files
In order to have a better structure, as the structs will get
sort functions and other methods
