+++ connectors/issues.go
@@ -1,0 +1,32 @@
+/*
+   Copyright 2017 Artem Sidorenko <artem@posteo.de>
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package connectors
+
+import (
+	"time"
+)
+
+// Issue describes an issue in the bug tracker
+type Issue struct {
+	ID         int
+	Name       string
+	ClosedDate time.Time
+	URL        string
+}
+
+// Issues is a slice with Issue elements
+type Issues []Issue
